requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
flask-jwt-extended==4.5.2
psycopg2-binary==2.9.7
redis==4.6.0
python-dotenv==1.0.0
requests==2.31.0
python-nmap==0.7.1
scapy==2.5.0
schedule==1.2.0
uuid==1.30